chore: pull helm chart upon release to remove duplication
To reduce confusion between public helm chart and the internal one,
we'll update the release workflow to pull the version specified in
version.txt from artifactory and use its NOTES.txt file for the release
notes.

Additionally, this removes the dependency on the chart-releaser-action,
which had some rough edges that required explicitly ignoring error
conditions and manually push gh-pages. This also obviates the need to
download previous releases to generate the index.yaml file (this was
going to be increasingly slower, and was already hitting GitHub API
limits).

Instead, we will manually pull the helm chart from artifactory, take the
previous index.yaml from gh-pages, and merge it with the newest
contents.